ai/requirements.txt
fastapi
pillow
torch
transformers
uvicorn
python-multipart
pydub
